moved product.service into product/
