+++ client/src/app/product/product-detail.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit }        from '@angular/core';
 import { ActivatedRoute, Params }   from '@angular/router';
-import { ProductService }           from '../product.service';
+import { ProductService }           from './product.service';
 
 @Component({
   selector: 'app-product-detail',
