+++ client/src/app/product/product.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterModule } from '@angular/router';
 
 import { ProductComponent } from './product.component';
-import { ProductService } from '../product.service';
+import { ProductService } from './product.service';
 
 describe('ProductComponent', () => {
   let component: ProductComponent;
