+++ client/src/app/product/product.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ProductService } from '../product.service';
+import { ProductService } from './product.service';
 
 @Component({
   selector: 'app-product',
